requirements.txt
appdirs==1.4.3
Django==1.10.1
django-extensions==1.7.9
django-tastypie==0.13.3
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
python-mimeparse==1.6.0
pytz==2016.6.1
six==1.10.0
virtualenv==15.1.0
